# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dictionary App API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/oauth/token": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "Token for access Apis",
                "operationId": "token",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/schemas/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Notfound"
                    },
                    "500": {
                        "$ref": "#/components/schemas/Forbidden"
                    }
                }
            }
        },
        "/search/suggestions": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Available Suggestions for keyword",
                "operationId": "suggestions",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/schemas/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Notfound"
                    },
                    "500": {
                        "$ref": "#/components/schemas/Forbidden"
                    }
                }
            }
        },
        "/search/keywords": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get Keywords",
                "operationId": "keywords",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/schemas/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Notfound"
                    },
                    "500": {
                        "$ref": "#/components/schemas/Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BadRequest": {
                "title": "Bad Request",
                "description": "BadRequest",
                "properties": {
                    "dsfsfds": {
                        "title": "dsfsfds",
                        "description": "Bad Request",
                        "type": "string"
                    }
                },
                "type": "string"
            },
            "Forbidden": {
                "title": "Forbidden",
                "description": "Forbidden",
                "properties": {
                    "forbidden": {
                        "title": "forbidden",
                        "description": "Forbidden",
                        "type": "string"
                    }
                },
                "type": "string"
            },
            "Notfound": {
                "title": "Not found",
                "description": "Not found",
                "properties": {
                    "notfound": {
                        "title": "Not found",
                        "description": "Not found",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "TokenRequest": {
                "title": "Create API acess token",
                "description": "Request for create a new user",
                "properties": {
                    "client_id": {
                        "title": "client_id",
                        "description": "client_id of key for storring",
                        "type": "string",
                        "example": "2"
                    },
                    "client_secret": {
                        "title": "client_secret",
                        "description": "client_secret for storring",
                        "type": "string",
                        "example": "8BSSg7qMYw2NAJaiMhQOCYxGlFSs141SLfPRLU"
                    }
                },
                "type": "object"
            },
            "TokenResponse": {
                "title": "token response",
                "description": "token response",
                "properties": {
                    "data": {
                        "type": "object",
                        "example": {
                            "client_id": "2",
                            "client_secret": "eyJ0eXAiOiJKVhJkpDuW_NQgUE4ccUTXVKXWrHhYL2Q88ncLi"
                        }
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "title": "Unauthenticated",
                "description": "Unauthenticated",
                "properties": {
                    "unauthenticated": {
                        "title": "Unauthenticated",
                        "description": "Unauthenticated",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            },
            "Token": {
                "type": "apiKey",
                "name": "Token",
                "in": "header",
                "scheme": "Token"
            }
        }
    },
    "tags": [
        {
            "name": "Dictionary App",
            "description": "API Endpoints of Dictionary App"
        }
    ],
    "security": {
        "0": [],
        "passport": {
            "type": "oauth2",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://localhost/oauth/authorize",
                    "tokenUrl": "http://localhost/oauth/token",
                    "refreshUrl": "http://localhost/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}